moved capabilities inside factories
